deprecate outdated listing logic and move button to styled components
